Inline leaf route logic into page.tsx files
Colocate welcome and unsubscribe UI in their App Router pages, export
Page as the default, and remove the extracted routes/ components.
