+++ apps/antalmanac/src/app/feedback/page.tsx
@@ -1,6 +1,6 @@
 import { FEEDBACK_LINK } from '$src/globals';
 import { redirect } from 'next/navigation';
 
-export default function FeedbackPage() {
+export default function Page() {
     redirect(FEEDBACK_LINK);
 }
