+++ apps/antalmanac/src/app/welcome/page.tsx
@@ -1,3 +1,17 @@
-import { NewUserPage } from '$routes/NewUserPage';
+'use client';
 
-export default NewUserPage;
+import { useSessionStore } from '$stores/SessionStore';
+import { useRouter } from 'next/navigation';
+import { useEffect } from 'react';
+
+export default function Page() {
+    const router = useRouter();
+    const setIsNewUser = useSessionStore((state) => state.setIsNewUser);
+
+    useEffect(() => {
+        setIsNewUser(true);
+        router.replace('/');
+    }, [router, setIsNewUser]);
+
+    return null;
+}
